Add files via upload
This paper analyses people's opinions online about wine consumption. We use data scraped through Vivinos website, the world largest wine marketplace. We compare the different characteristics of 13974 wines such as the country, grape type, type of wine, price and alcohol percentage. All in order to see if we can predict the wine rating based on these attributes. These ratings are from 1 to 5. We apply Ridge and Lasso, two different machine learning models with different polynomials degrees, to predict wine rating. Although our model was not able to predict the behaviour of the users correctly, based on these features, for wines rated above 3.5, the model performed better - but still was not sufficient to make a powerful assumption about the reviews it will describe.